agents/CICD_agent/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
requests
python-dotenv
